:recycle: refactor: 속성 수정

+++ backend/src/main/java/com/osakak/jusangnakwon/domain/liquor/entity/Cocktail.java
@@ -43,7 +43,7 @@ public class Cocktail {
     private Object scrap_users;
     // RatingUser를 객체로 사용
     private Object ratings;
-    private int rating_avg;
+    private Double rating_avg;
     private List<String> feed_id;
 
    /* public static CocktailBuilder builder(Cocktail cocktail) {
