+++ backend/src/main/java/com/osakak/jusangnakwon/domain/liquor/entity/TraditionalLiquor.java
@@ -31,7 +31,7 @@ public class TraditionalLiquor {
     private Object scrap_users;
     // RatingUser를 객체로 사용
     private Object ratings;
-    private int rating_avg;
+    private Double rating_avg;
     private List<String> feed_id;
 
 }
